+++ src/core/enums/index.ts
@@ -1,3 +1,2 @@
-export * from './role.enum'
 export * from './action.enum'
 export * from './user.enum'
