--- src/core/enums/role.enum.ts
@@ -1,7 +1,0 @@
-
-export enum ROLE {
-  ADMIN = 'Admin',
-  LAWYER = 'Lawyer',
-  USER = 'User',
-  CLIENT = 'Client',
-}
