+++ src/core/enums/user.enum.ts
@@ -3,6 +3,21 @@ export enum GENDER {
   FEMALE = 'female',
 }
 
+export enum ROLE {
+  ADMIN = 'Admin',
+  LAWYER = 'Lawyer',
+  USER = 'User',
+  CLIENT = 'Client',
+}
+
+export enum STATUS {
+  NONE = '',
+  ACTIVE = 'Active',
+  REJECT = 'Reject',
+  PENDING = 'Pending',
+  BLOCK = 'Block',
+}
+
 export enum SPECIALIZATION {
   CIVIL = 'Civil',
   CRIMINAL = 'Criminal',
@@ -13,11 +28,6 @@ export enum COURT {
   LOWER = 'Lower Court',
   HIGH = 'High Court',
   SUPREME = 'Supreme Court',
+
 }
 
-export enum STATUS {
-  ACTIVE = 'Active',
-  REJECT = 'Reject',
-  PENDING = 'Pending',
-  BLOCK = 'Block',
-}
